#210 Clean up google_calendar_backend
